Remove component recycling
