+++ src/vdom/component-utils.js
@@ -1,0 +1,24 @@
+import { Component } from '../component';
+
+/** Create a component. Normalizes differences between PFC's and classful Components. */
+export function createComponent(Ctor, props, context) {
+	let inst;
+
+	if (Ctor.prototype && Ctor.prototype.render) {
+		inst = new Ctor(props, context);
+		Component.call(inst, props, context);
+	}
+	else {
+		inst = new Component(props, context);
+		inst.constructor = Ctor;
+		inst.render = doRender;
+	}
+
+	return inst;
+}
+
+
+/** The `.render()` method for a PFC backing instance. */
+function doRender(props, state, context) {
+	return this.constructor(props, context);
+}
